fix: useReactionFastenings selector
