+++ src/chats/components/conversation/ScrollButton.tsx
@@ -12,7 +12,7 @@ import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
 import useEventListener, { EventName, NewMessageEvent } from '../../../hooks/useEventListener';
-import { getRoomUnreadsSelector } from '../../../store/selectors/ChatsRegistrySelectors';
+import { getRoomUnreadSelector } from '../../../store/selectors/ChatsRegistrySelectors';
 import { getRoomMutedSelector } from '../../../store/selectors/RoomsSelectors';
 import { getUserId } from '../../../store/selectors/SessionSelectors';
 import useStore from '../../../store/Store';
@@ -40,7 +40,7 @@ export const ScrollBadge = styled(Badge)`
 `;
 
 const ScrollButton = ({ roomId, onClickCb }: ScrollButtonProps): ReactElement => {
-	const unreadCount = useStore((store) => getRoomUnreadsSelector(store, roomId));
+	const unreadCount = useStore((store) => getRoomUnreadSelector(store, roomId));
 	const roomMuted = useStore((state) => getRoomMutedSelector(state, roomId));
 	const myUserId = useStore(getUserId);
