+++ src/store/slices/useChatsRegistryStoreSlice.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-param-reassign */
 /*
- * SPDX-FileCopyrightText: 2022 Zextras <https://www.zextras.com>
+ * SPDX-FileCopyrightText: 2025 Zextras <https://www.zextras.com>
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
